requirements.txt
# Core HTTP + parsing
requests==2.32.3            # Robust HTTP client
beautifulsoup4==4.12.3      # HTML parsing
lxml==5.2.2                 # Fast parser for BeautifulSoup
tqdm==4.66.4                # Progress bars

# Web Framework
Flask==3.0.0                # Web application framework
Werkzeug==3.0.1             # WSGI utilities
gunicorn>=21.2.0,<22.0.0    # Production WSGI server

# Data handling / utilities
pandas>=2.2.2               # Optional tabular transforms (flexible version)
python-dateutil==2.9.0.post0

# Embeddings + vector store
sentence-transformers>=2.2.2,<3.0.0  # Embedding models (compatible range)
chromadb==1.0.15              # Persistent vector DB - FIXED VERSION

# LangChain toolkit (updated to latest compatible versions)
langchain>=0.2.11,<0.4.0
langchain-community>=0.2.10,<0.4.0  
langchain-text-splitters>=0.2.2,<0.4.0

# API
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0

# Settings/validation
pydantic>=2.8.2,<3.0.0
pydantic-settings>=2.3.4,<3.0.0

# Async/HTTP client (optional helpers)
httpx>=0.27.0,<1.0.0

# Testing
pytest>=8.2.2,<9.0.0

# Env loading
python-dotenv>=1.0.1,<2.0.0

# Reporting / documentation generation
python-docx==1.1.2

